Added TextureManager.

+++ src/main/java/org/halvors/Game/Client/Location.java
@@ -2,13 +2,13 @@ package org.halvors.Game.Client;
 
 public class Location {
 	private World world;
-	private double x;
-	private double y;
-	private double z;
+	private float x;
+	private float y;
+	private float z;
 	private float pitch;
     private float yaw;
 	
-	public Location(World world, double x, double y, double z, float pitch, float yaw) {
+	public Location(World world, float x, float y, float z, float pitch, float yaw) {
 		setWorld(world);
 		setX(x);
 		setY(y);
@@ -25,27 +25,27 @@ public class Location {
 		this.world = world;
 	}
 
-	public double getX() {
+	public float getX() {
 		return x;
 	}
 
-	public void setX(double x) {
+	public void setX(float x) {
 		this.x = x;
 	}
 
-	public double getY() {
+	public float getY() {
 		return y;
 	}
 
-	public void setY(double y) {
+	public void setY(float y) {
 		this.y = y;
 	}
 
-	public double getZ() {
+	public float getZ() {
 		return z;
 	}
 
-	public void setZ(double z) {
+	public void setZ(float z) {
 		this.z = z;
 	}
